requirements.txt
pymongo
python-dotenv
streamlit
bcrypt